refactor: move docs closer to what it's doc'ing
